Add support for (remote) pigpio
Selecting pigpio as pinfactory for gpiozero allows using remote gpio,
for example on a different raspberry. This is especially usefull when
the docker container is running on a host that is not location where you
want to use the gpio, for example to measure watermeter pulses.

Adding GPIOZERO_PIN_FACTORY=pigpio and PIGPIO_ADDR=<somehost> to the
docker-compose file is sufficient to allow p1monitor to use the gpios of
a remote raspberry. The remote raspberry has to run the pigpiod.service,
which is available in the pigpio raspbian/debian package.